ci: do not rebuild images before test (#131)
rmf image is always rebuilt because base layers of multi-stage build are discarded, we can fix it by going back to pre-multistage support by separating each stage into individual docker images and have rmf depend on them.

But doing so is probably not worth it, the only time rebuilding images before tests is useful is when a PR requires some changes in the dockerfiles. Hopefully this is will only happen very rarely.