ci: run e2e ci in container (#151)
Previously there is a difference between running e2e in CI and in local runs. The CI run launches a docker container and map the repo onto it, this caused problems with permissions and random failures where it works on the dev's PC but fails in CI.

This PR changed it so that the CI is ran using the e2e container, allowing us to use github actions on it. This also allows the CI run to use the same commands as local runs, removing the environment difference.