feat(api-server): door health monitoring (#233)
This adds a watchdog for doors, currently the only check is a heartbeat on the door state.

There is also a change in logging/database schema, now we only keep the working set in the database, logs are printed to stdout to be collected by another service. Presumably the logs will be collected in a separate database optimized for writes, so the sql database will just be used for the current/latest states.