fix(example-deployment): fix reporting infinite redirect (#460)
The root cause is because `/etc/nginx/conf.d/default.conf` already contains a `location / {}` block and that conflicts with the declaration in `/etc/nginx/conf.d/reporting.conf`. The reason it works in dashboard is because `dashboard.conf` gets loaded before `default.conf` because or file system sorting ("da" comes before "de").

This also adds the admin user for use with the new authorization changes. It should be used in favor of "example" when performing some things like submitting tasks.

Signed-off-by: Teo Koon Peng <koonpeng@openrobotics.org>