perf: dashboard (#490)
New building map panel is unusable in larger maps (e.g. clinic) due to large number of entities. This workaround it by rendering at a fixed rate as opposed to whenever new data is available, rendering of each cell/row is also improved by the use of memos.

This is not an ideal solution, for the long term, we should use virtualized table/lists. Also this does not address the design issues when there are lots of entities.

Signed-off-by: Teo Koon Peng <koonpeng@openrobotics.org>