build: switch to openapi generator (#534)
The openapi generator seems to be more updated, and more importantly, it does not have the typing bugs in the typescript-axios generator. Since there have been no movement on the PRs to fix typings in the swagger generator, this PR switches to the community generator.

This also attempts to use types from api-client, which contains the types from the api-server (instead of from ROS) whenever possible. There are still many places which uses rmf-models because of the lack of enum support (related open-rmf/rmf#108).

Overall, this should improve the type checkings as there will be no need to perform type conversions at many components. In the future, when RMF supports a web-based public api, we should also remove all the legacy generated ros typings, so this serves as a stepping stone for the future work.

Signed-off-by: Teo Koon Peng <koonpeng@openrobotics.org>