refactor: backport tasksv2 architecture changes (#565)
The problem is that the whole fastapi stack relies on singletons, the api-server tries to avoid them but the new tasksv2 system requires us to run multiple servers under the same process making this avoidance of singletons much harder. This PR changes the api-server to embrace the singleton patterns of fastapi.

Some tests like tasks routes and permissions have to be removed because it uses mocks which are hard to set up in the new architecture, the current tasks system will be replaced anyway so there is not much point in making the mocks work. The permissions tests are technically not needed if permissions are enforced throughout the app (the normal get/post tests should cover authorisations as well), it is not done so now because it is lacking rmf support.

Signed-off-by: Teo Koon Peng <koonpeng@openrobotics.org>

Co-authored-by: ChawinTan <chawin15@gmail.com>